move loadDefinition
