+++ examples/cli/cli.spec.js
@@ -1,22 +1,25 @@
 import { promises as fs } from 'fs';
 import { createRequire } from 'module';
-import { dirname } from 'path';
+import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
-import { promisify } from 'util';
-import { exec } from 'child_process';
+import { spawnSync } from 'child_process';
+import { loadDefinition } from './utils.js';
 
-const sh = promisify(exec);
 const require = createRequire(import.meta.url);
 const __dirname = dirname(fileURLToPath(import.meta.url));
-const bin = `node ${__dirname}/cli.js`;
+const cli = `${__dirname}/cli.js`;
 
 describe('Example command line application', () => {
-  it('should produce results matching reference specification', async () => {
-    const result = await sh(
-      `${bin} --definition test/fixtures/swaggerDefinition/example.js --apis examples/app/parameters.* examples/app/route*`
-    );
-    expect(result.stderr).toBe('');
-    expect(result.stdout).toBe(
+  it('should produce results matching reference specification', () => {
+    const { stderr, stdout } = spawnSync(cli, [
+      '--definition',
+      'examples/swaggerDefinition/example.js',
+      '--apis',
+      'examples/app/parameters.*',
+      'examples/app/route*',
+    ]);
+    expect(stderr.toString()).toBe('');
+    expect(stdout.toString()).toBe(
       'Specification has been created successfully!\n'
     );
     const refSpec = require('./reference-specification.json');
@@ -26,5 +29,75 @@ describe('Example command line application', () => {
 
   afterAll(async () => {
     await fs.unlink(`${process.cwd()}/swagger.json`);
+  });
+});
+
+describe('loadDefinition', () => {
+  const example = '../swaggerDefinition/example';
+
+  it('should throw on bad input', async () => {
+    await expect(loadDefinition('bad/path/to/nowhere')).rejects.toThrow(
+      'Definition file should be any of the following: .js, .mjs, .cjs, .json, .yml, .yaml'
+    );
+  });
+
+  it('should support .json', async () => {
+    const def = resolve(__dirname, `${example}.json`);
+    const result = await loadDefinition(def);
+    expect(result).toEqual({
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+        description: 'A sample API',
+      },
+    });
+  });
+
+  it('should support .yaml', async () => {
+    const def = resolve(__dirname, `${example}.yaml`);
+    const result = await loadDefinition(def);
+    expect(result).toEqual({
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+        description: 'A sample API',
+      },
+    });
+  });
+
+  it('should support .js', async () => {
+    const def = resolve(__dirname, `${example}.js`);
+    const result = await loadDefinition(def);
+    expect(result).toEqual({
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+        description: 'A sample API',
+      },
+    });
+  });
+
+  it('should support .cjs', async () => {
+    const def = resolve(__dirname, `${example}.cjs`);
+    const result = await loadDefinition(def);
+    expect(result).toEqual({
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+        description: 'A sample API',
+      },
+    });
+  });
+
+  it('should support .mjs', async () => {
+    const def = resolve(__dirname, `${example}.mjs`);
+    const result = await loadDefinition(def);
+    expect(result).toEqual({
+      info: {
+        title: 'Hello World',
+        version: '1.0.0',
+        description: 'A sample API',
+      },
+    });
   });
 });
