+++ examples/swaggerDefinition/example.yaml
